@@ -13,21 +13,21 @@
         "cloudinary": "^1.29.0",
         "cors": "^2.8.5",
         "multer": "^1.4.4",
-        "overlay-navbar": "^1.1.1",
-        "react": "^17.0.2",
+        "overlay-navbar": "^1.2.3",
+        "react": "^18.0.0",
         "react-alert": "^7.0.3",
         "react-alert-template-basic": "^1.0.2",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.0.0",
         "react-helmet": "^6.1.0",
         "react-icons": "^3.11.0",
-        "react-js-pagination": "^3.0.3",
+        "react-js-pagination": "^3.0.2",
         "react-material-ui-carousel": "^3.2.0",
         "react-paginate": "^8.1.2",
         "react-rating-stars-component": "^2.2.0",
         "react-redux": "^7.2.6",
         "react-router": "^6.2.2",
         "react-router-dom": "^6.2.2",
-        "react-scripts": "^2.1.3",
+        "react-scripts": "^5.0.0",
         "redux": "^4.1.2",
         "redux-devtools-extension": "^2.13.9",
         "redux-thunk": "^2.4.1",
