@@ -1,10 +1,3 @@
-/***
- * @Author:xxx
- * @Date:2022-02-26 23:44:51
- * @LastModifiedBy:xxx
- * @Last Modified time:2022-02-26 23:44:52
- */
-
 const mongoose = require('mongoose');
 const validator = require('validator');
 const bcrypt = require('bcryptjs');
@@ -12,81 +5,85 @@ const jwt = require('jsonwebtoken');
 const crypto = require('crypto');
 
 const userSchema = new mongoose.Schema({
-   name: {
-      type: String,
-      required: [true, 'Please Enter Your Name'],
-      maxLength: [30, 'Name cannot exceed 30 characters'],
-      minLength: [4, 'Name must be atleast of 4 characters long'],
-   },
-   email: {
-      type: String,
-      required: [true, 'Please Enter Your Email'],
-      unique: true,
-      validate: [validator.isEmail, 'Please Enter a valid Email'],
-   },
-   password: {
-      type: String,
-      required: [true, 'Please Enter Your Password'],
-      minLength: [4, 'Password must be atleast of 8 characters long'],
-      select: false,
-   },
-   avatar: {
-      public_id: {
-         type: String,
-         required: true,
-      },
-      url: {
-         type: String,
-         required: true,
-      },
-   },
-   role: {
-      type: String,
-      default: 'user',
-   },
-   resetPasswordToken: String,
-   resetPasswordExpire: Date,
+    name: {
+        type: String,
+        required: [true, 'Please Enter Your Name'],
+        maxLength: [30, 'Name cannot exceed 30 characters'],
+        minLength: [4, 'Name must be atleast of 4 characters long'],
+    },
+    email: {
+        type: String,
+        required: [true, 'Please Enter Your Email'],
+        unique: true,
+        validate: [validator.isEmail, 'Please Enter a valid Email'],
+    },
+    password: {
+        type: String,
+        required: [true, 'Please Enter Your Password'],
+        minLength: [4, 'Password must be atleast of 8 characters long'],
+        select: false,
+    },
+    avatar: {
+        public_id: {
+            type: String,
+            required: true,
+        },
+        url: {
+            type: String,
+            required: true,
+        },
+    },
+    role: {
+        type: String,
+        default: 'user',
+    },
+    createdAt: {
+        type: Date,
+        default: Date.now,
+    },
+    resetPasswordToken: String,
+    resetPasswordExpire: Date,
 });
 
 userSchema.pre('save', async function (next) {
-   if (!this.isModified('password')) {
-      next();
-   }
+    if (!this.isModified('password')) {
+        next();
+    }
 
-   this.password = await bcrypt.hash(this.password, 12);
+    this.password = await bcrypt.hash(this.password, 12);
 });
 
 // jwt token
 userSchema.methods.getJWTToken = function () {
-   return jwt.sign(
-      {
-         id: this._id,
-      },
-      process.env.JWT_SECRET_KEY,
-      {
-         expiresIn: process.env.JWT_EXPIRES_IN,
-      }
-   );
+    return jwt.sign(
+        {
+            id: this._id,
+        },
+        process.env.JWT_SECRET_KEY,
+        {
+            expiresIn: process.env.JWT_EXPIRES_IN,
+        }
+    );
 };
 
 // compare Password
 userSchema.methods.comparePassword = async function (enteredPassword) {
-   return await bcrypt.compare(enteredPassword, this.password);
+    return await bcrypt.compare(enteredPassword, this.password);
 };
 
 // generating password reset token
 userSchema.methods.getResetPasswordToken = function () {
-   // generating token
-   const resetToken = crypto.randomBytes(20).toString('hex');
+    // generating token
+    const resetToken = crypto.randomBytes(20).toString('hex');
 
-   // hashing and add to userSchema
-   this.resetPasswordToken = crypto
-      .createHash('sha256')
-      .update(resetToken)
-    .digest('hex');
-  this.resetPasswordExpire = Date.now() + 15 * 60 * 1000;
+    // hashing and add to userSchema
+    this.resetPasswordToken = crypto
+        .createHash('sha256')
+        .update(resetToken)
+        .digest('hex');
+    this.resetPasswordExpire = Date.now() + 15 * 60 * 1000;
 
-  return resetToken;
+    return resetToken;
 };
 
 module.exports = mongoose.model('User', userSchema);
