@@ -1,3 +1,10 @@
+/***
+ * @Author:xxx
+ * @Date:2022-04-13 23:50:44
+ * @LastModifiedBy:xxx
+ * @Last Modified time:2022-04-13 23:50:46
+ */
+
 import axios from 'axios';
 
 import {
@@ -13,6 +20,9 @@ import {
     LOAD_USER_FAIL,
     LOGOUT_SUCCESS,
     LOGOUT_FAIL,
+    UPDATE_PROFILE_REQUEST,
+    UPDATE_PROFILE_SUCCESS,
+    UPDATE_PROFILE_FAIL,
 } from '../constants/userConstants';
 
 // Login
@@ -20,8 +30,12 @@ export const login = (email, password) => async (dispatch) => {
     try {
         dispatch({ type: LOGIN_REQUEST });
 
-        const config = { headers: { "Content-Type" : "application/json" } };
-        const { data } = await axios.post(`/api/v1/login`, { email, password }, config);
+        const config = { headers: { 'Content-Type': 'application/json' } };
+        const { data } = await axios.post(
+            `/api/v1/login`,
+            { email, password },
+            config
+        );
         dispatch({ type: LOGIN_SUCCESS, payload: data.user });
     } catch (error) {
         dispatch({ type: LOGIN_FAIL, payload: error.response.data.message });
@@ -37,8 +51,9 @@ export const register = (userData) => async (dispatch) => {
             headers: { 'Content-Type': 'multipart/form-data' },
         };
 
-        console.log({ userData });
-        const { data } = await axios.post(`/api/v1/register`, userData, { config });
+        const { data } = await axios.post(`/api/v1/register`, userData, {
+            config,
+        });
 
         dispatch({ type: REGISTER_USER_SUCCESS, payload: data.user });
     } catch (error) {
@@ -49,16 +64,14 @@ export const register = (userData) => async (dispatch) => {
     }
 };
 
-// load user 
+// load user
 export const loadUser = () => async (dispatch) => {
     try {
         dispatch({
             type: LOAD_USER_REQUEST,
         });
 
-        const { data } = await axios.get(
-            `/api/v1/me`,
-        );
+        const { data } = await axios.get(`/api/v1/me`);
         dispatch({ type: LOAD_USER_SUCCESS, payload: data.user });
     } catch (error) {
         dispatch({
@@ -71,7 +84,6 @@ export const loadUser = () => async (dispatch) => {
 // logout User
 export const logout = () => async (dispatch) => {
     try {
-
         await axios.get(`/api/v1/logout`);
         dispatch({ type: LOGOUT_SUCCESS });
     } catch (error) {
@@ -82,6 +94,25 @@ export const logout = () => async (dispatch) => {
     }
 };
 
+// Update Profile
+export const updateProfile = (userData) => async (dispatch) => {
+    try {
+        dispatch({ type: UPDATE_PROFILE_REQUEST });
+
+        const config = { headers: { 'Content-Type': 'multipart/form-data' } };
+
+        const { data } = await axios.put(`/api/v1/me/update`, userData,
+            config,
+        );
+
+        dispatch({ type: UPDATE_PROFILE_SUCCESS, payload: data.success });
+    } catch (error) {
+        dispatch({
+            type: UPDATE_PROFILE_FAIL,
+            payload: error.response.data.message,
+        });
+    }
+};
 
 // clearing errors
 export const clearErrors = () => async (dispatch) => {
