@@ -1,11 +1,11 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import { productDetailsReducer, productReducer } from './reducers/productReducer';
+import { productDetailsReducer, productsReducer } from './reducers/productReducer';
 import { userReducer } from './reducers/userReducer';
 
 const reducer = combineReducers({
-    products: productReducer,
+    products: productsReducer,
     productDetails: productDetailsReducer,
     user: userReducer
 });
